Bubble buttons in seperate component
