--- app/frontend/src/components/AssistantBubble.css
@@ -1,34 +1,0 @@
-@keyframes blink {  
-    0% {opacity: 1;}  
-    50% {opacity: 0;}  
-    100% {opacity: 1;}  
-  }  
-    
-.blink {  
-    animation: blink 1s linear infinite;  
-    font-size: x-small;
-}  
-    
-  .blink:nth-child(2) {  
-    animation-delay: 0.2s;  
-  }  
-    
-  .blink:nth-child(3) {  
-    animation-delay: 0.4s;  
-  }  
-  
-  .copyButton {  
-    cursor: pointer;  
-    background-color: transparent;  
-    border: none;  
-  }  
-  
-  .container {  
-    display: flex;  
-    align-items: center;  
-    justify-content: space-between;  
-  }  
-    
-  .text {  
-    flex-grow: 1;  
-  }
